created multiple testing classes

+++ src/main/java/language/backend/compiler/bytecode/Compiler.java
@@ -1232,8 +1232,8 @@ public class Compiler {
 
     public final Map<String, String> missMatchMap = new HashMap<>() {{
 
-        put("f64", "i64");
-        put("i64", "f64");
+        put("f32", "i64");
+        put("i64", "f32");
         put("i32", "i64");
         put("l64", "i64");
         put("i16", "i64");
